+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## [1.0.0] - 2020-04-01
 ### Added
-- All Symfony Commands (78) by [@nalabdou](https://github.com/nalabdou).
+- All Symfony Commands by [@nalabdou](https://github.com/nalabdou).
+## [1.1.0] - 2021-27-04
+### Added
+- Add superConsole.path configration Now you can set run path like that docker-compose exec <container> bin/console instead php bin/console #5
+### Removed
+- Remove some deprecated commands #4